backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.135.1
h11==0.16.0
idna==3.11
joblib==1.5.3
MarkupSafe==3.0.3
motor==3.7.1
neo4j==5.28.1
numpy==2.4.3
pandas==3.0.1
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-http-client==3.3.7
python-jose==3.5.0
python-multipart==0.0.22
pytz==2026.1.post1
requests==2.32.5
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
sendgrid==6.12.5
six==1.17.0
starlette==0.52.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
Werkzeug==3.1.6
